// app/(employees)/(tabs)/types.ts
export interface ShiftDetailsProps {
  id: string;
    name: string;
    location: string;
    date: Date;
    background_image_url: string;
    company_name: string;
    salary_per_time: string;
    total_earning: number;
    shift_duration: string;
    location_map: string;
    start_time: Date;
    end_time: Date;
    requirement_items: any;
    disallowed_items: any;
    requirement_description: string;
    status: string;
    created_at: Date;
    updated_at: Date;
    employer: string;
}
// ...
  export interface OngoingShiftProps {
    id: string;
    actual_start_time: Date;
    actual_end_time: Date;
    shift_progress: string;
    created_at: Date;
    updated_at: Date;
    employee: string;
    employer: string;
    shift: string;
}